Add map project and layer introspection script
Added map project with appropriate data from the data catalogue for inspection and validation of relevant data. Besides validation of data integrity, it can be utilised as a project scaffold for bootstrapping further data analysis, providing all the available layers in a single project.

By default, all layers (other than basic layers required for locality purposes) should have their visibility disabled in order to improve project startup performance. Project is currently configured and designed to work with pg_service functionality.

Additionally includes an introspection python script for interrogating the available layers for source details and other attributes. Script commentary is verbose to assist new users with understanding and leveraging the pyqgis API for data auditing. Script is supplied as a text file as it is designed for utilisation from the python console within the QGIS Interface and is not intended as a standalone script.